refactor: 공통 컬러 테마 시스템 common 모듈로 이동 (#53)
Closes #53
